+++ packages/blocks/checkout-billing-payment/vitest.config.mjs
@@ -1,0 +1,3 @@
+import { config } from '@o2s/vitest-config/block';
+
+export default config;